+++ packages/core/src/binders/class-name.binder.ts
@@ -7,6 +7,8 @@ import { IBinder } from '../interfaces';
  * Adds a class (whatever value is in place of [classname]) on the element
  * when the value evaluates to true and removes that class if the value
  * evaluates to false.
+ * @example
+ * <li rv-class-completed="todo.done">{ todo.name }</li>
  */
 export const classStarBinder: IBinder<boolean> = {
   name: 'class-*',