+++ packages/core/src/formatters/math/random.formatter.ts
@@ -2,7 +2,7 @@ import { Utils } from '../../services/utils';
 
 /**
  * Get a back random value of array
- * @example <div rv-class='"["col-2", "col-3", "col-4", "col-5", "col-6"]" | parse | random'>
+ * @example <div rv-add-class='"["col-2", "col-3", "col-4", "col-5", "col-6"]" | parse | random'>
  */
 export const random = {
   name: 'random',