+++ packages/core/src/formatters/property/parse.formatter.ts
@@ -2,7 +2,7 @@ import { Utils } from '../../services/utils';
 
 /**
  * parse json string to object
- * @example <div rv-class='"["col-2", "col-3", "col-4", "col-5", "col-6"]" | parse | random'>
+ * @example <div rv-add-class='"["col-2", "col-3", "col-4", "col-5", "col-6"]" | parse | random'>
  */
 export const parse = {
   name: 'parse',